requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-filter==2.2.0
django-model-utils==3.2.0
djangorestframework==3.10.3
idna==2.8
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.5
xlrd==1.2.0
